--- new_task2_results/templates/new_task2_results/MyPage.html
@@ -1,13 +1,0 @@
-{% extends "global/Page.html" %}
-{% load otree static %}
-
-{% block title %}
-    Page title
-{% endblock %}
-
-{% block content %}
-
-    {% formfields %}
-    {% next_button %}
-
-{% endblock %}
