+++ settings.py
@@ -25,7 +25,7 @@ APP_KITS = {
             'task2_prac_tr13',
             'task2_quiz',
             'task2_tr1',
-            'new_task2_results',
+            'task2_results',
             'static_portfolio',
             'static_portfolio_results',
             'static_portfolio_b',
@@ -48,7 +48,7 @@ APP_KITS = {
             'task2_prac_tr2',
             'task2_quiz',
             'task2_tr2',
-            'new_task2_results',
+            'task2_results',
             'static_portfolio_chp',
             'static_portfolio_results',
             'static_portfolio_b_chp',
@@ -70,8 +70,8 @@ APP_KITS = {
             'task2_instructions',
             'task2_prac_tr13',
             'task2_quiz',
-            'new_task2',
-            'new_task2_results',
+            'task2_tr1',
+            'task2_results',
             'static_portfolio',
             'static_portfolio_results',
             'static_portfolio_b',
@@ -106,7 +106,7 @@ SESSION_CONFIGS = [
                       'task2_quiz',
                       'task2_tr1',
                       'task2_tr2',
-                      'new_task2_results',
+                      'task2_results',
                       'static_portfolio',
                       'static_portfolio_chp',
                       'static_portfolio_results',
