+++ task2_results/models.py
@@ -12,7 +12,7 @@ Your app description
 
 
 class Constants(BaseConstants):
-    name_in_url = 'new_task2_results'
+    name_in_url = 'task2_results'
     players_per_group = None
     num_rounds = 1
     initial_wealth = 100
